doing some cleanup, reorganization, and creating dynamic outputs for each image handler row!
