+++ js/Timeline-UI.js
@@ -1,12 +1,7 @@
 import { app } from "../../scripts/app.js";
 import './utils/Sortable.min.js';
-import { makeNodeManager, out } from "./managers/NodeManager.js";
-
-const linkElem = document.createElement("link");
-linkElem.rel = "stylesheet";
-linkElem.href = "./timeline-styles.css";
-// document.head.appendChild(linkElem);
-
+import { makeNodeManager } from "./managers/NodeManager.js";
+import { loadCSS } from "./utils/MiscUtils.js";
 
 function sendDataToBackend(image_timelines) {
     let data = {imageData: []};
@@ -73,5 +68,5 @@ const node = {
   },
 };
 
-
+loadCSS();
 app.registerExtension(node);
